+++ examples/wrap-example/index.js
@@ -7,7 +7,10 @@ Vue.use(VueLoading);
 
 new Vue({
   el: '#app',
-  vueLoading: new VueLoading({ registerComponents: false }),
+  vueLoading: new VueLoading({
+    registerComponents: false,
+    componentName: 'my-spinner'
+  }),
   render: function(createElement) {
     return createElement(main);
   }